+++ public/functions-shortcode-filters.php
@@ -67,7 +67,7 @@ function arve_sc_filter_attr( $a ) {
 			$iframe_src .= '#t=' . (int) $a['start'];
 		}
 
-		$iframe_sandbox = 'allow-scripts allow-same-origin allow-popups';
+		$iframe_sandbox = 'allow-scripts allow-same-origin allow-presentation allow-popups';
 
 		if ( 'vimeo' == $a['provider'] ) {
 			$iframe_sandbox .= ' allow-forms';